+++ src/components/GameTable.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import ReactFire from 'reactfire';
 import Firebase from 'firebase';
-import Player from './player';
-import AddPlayerForm from './player-form';
-import Icon from './icon';
+import Player from './gametable/player';
+import AddPlayerForm from './gametable/player-form';
+import Icon from './common/icon';
 import conf from '../../app.config.json';
 import _ from 'lodash';
 import Elo from 'elo-rank';
