+++ src/components/gametable/play-buttons.jsx
