+++ src/components/gametable/player-form.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import _validate from '../utils/validator'
+import _validate from '../../utils/validator'
 import _ from 'lodash';
 
 module.exports = React.createClass( {
